add scale100.js
